arkos: convert msx examples into single files (classic)
