--- libsrc/psg/ay/examples/arkos/msx/interrupt-rom/Makefile
@@ -1,20 +1,0 @@
-# main target objs
-MAIN_OBJS = main.o hocuspocus.o
-
-ZCC = zcc +msx -subtype=rom -vn
-CFLAGS = -DARKOS_USE_ROM_PLAYER
-
-# targets
-.PHONY: build clean
-
-build: $(MAIN_OBJS)
-	$(ZCC) -m -s --list -create-app -o arkos $(MAIN_OBJS)
-
-%.o: %.c
-	$(ZCC) $(CFLAGS) -c -s --list $*.c
-
-%.o: %.asm
-	$(ZCC) $(CFLAGS) -c -s --list $*.asm
-
-clean:
-	@-rm *.lis *.bin *.tap *.o *.sym *.map *.save arkos *.rom 2>/dev/null
